+++ src/models/Project.ts
@@ -37,8 +37,8 @@ export default class Project extends Model {
         join: {
           from: 'project.id',
           through: {
-            from: 'projectmembership.projectId',
-            to: 'projectmembership.userId',
+            from: 'projectMembership.projectId',
+            to: 'projectMembership.userId',
             extra: ['membershipType', 'membershipIsActive']
           },
           to: 'user.id'